backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
werkzeug
requests
waitress
numpy>=1.20
Pillow>=9.0.0
mediapipe>=0.10.14
rembg>=2.0.53
onnxruntime>=1.16.0
scipy>=1.10.0
scikit-learn>=1.2.0
curl_cffi>=0.6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
gradio-client>=0.7.0
requests>=2.28.0
python-dotenv
